test(use-pan-event): migrate interaction tests to browser mode
